+++ pkg/client/types/file.go
@@ -56,8 +56,8 @@ const (
 	GroupRoundRobin    WorkflowConcurrencyLimitStrategy = "GROUP_ROUND_ROBIN"
 	DropNewest         WorkflowConcurrencyLimitStrategy = "DROP_NEWEST"
 	QueueNewest        WorkflowConcurrencyLimitStrategy = "QUEUE_NEWEST"
-	CancelExceptNewest WorkflowConcurrencyLimitStrategy = "CANCEL_EXCEPT_NEWEST"
-	CancelExceptOldest WorkflowConcurrencyLimitStrategy = "CANCEL_EXCEPT_OLDEST"
+	CancelQueuedExceptNewest WorkflowConcurrencyLimitStrategy = "CANCEL_QUEUED_EXCEPT_NEWEST"
+	CancelQueuedExceptOldest WorkflowConcurrencyLimitStrategy = "CANCEL_QUEUED_EXCEPT_OLDEST"
 )
 
 type WorkflowConcurrency struct {