+++ pkg/repository/workflow.go
@@ -81,7 +81,7 @@ type CreateConcurrencyOpts struct {
 	MaxRuns *int32 `json:"maxRuns,omitempty" validate:"omitempty,min=1"`
 
 	// (optional) the strategy to use when the concurrency limit is reached, default CANCEL_IN_PROGRESS
-	LimitStrategy *string `validate:"omitnil,oneof=CANCEL_IN_PROGRESS GROUP_ROUND_ROBIN CANCEL_NEWEST CANCEL_EXCEPT_NEWEST CANCEL_EXCEPT_OLDEST"`
+	LimitStrategy *string `validate:"omitnil,oneof=CANCEL_IN_PROGRESS GROUP_ROUND_ROBIN CANCEL_NEWEST CANCEL_QUEUED_EXCEPT_NEWEST CANCEL_QUEUED_EXCEPT_OLDEST"`
 
 	// (required) a concurrency expression for evaluating the concurrency key
 	Expression string `validate:"celworkflowrunstr"`