+++ sdks/ruby/src/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Adds support for `CANCEL_EXCEPT_NEWEST` and `CANCEL_EXCEPT_OLDEST` concurrency strategies.
+- Adds support for `CANCEL_QUEUED_EXCEPT_NEWEST` and `CANCEL_QUEUED_EXCEPT_OLDEST` concurrency strategies.
 
 
 ## [0.6.0] - 2026-07-23